requirements.txt
faiss_cpu==1.7.4
fastapi==0.104.1
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
pandas==2.1.2
Pillow==9.5.0
Pillow==10.1.0
Requests==2.31.0
torch==2.1.0+cu118
torchvision==0.15.2+cu118
uvicorn==0.24.0.post1
